requirements.txt
watchdog
langchain
openai
python-dotenv